move download_url_to_file to downloader.py
